+++ Api/Api.lua
@@ -1,9 +1,9 @@
-local _, addon = ...;
-KrowiEVU = {};
+local _, addon = ...
+KrowiEVU = {}
 
-addon.Api = {};
-local api = addon.Api;
+addon.Api = {}
+local api = addon.Api
 
 function api.Load()
-	KrowiEVU_MerchantItemsContainer = addon.Gui.MerchantItemsContainer; -- Vendorer uses this
+	KrowiEVU_MerchantItemsContainer = addon.Gui.MerchantItemsContainer -- Vendorer uses this
 end