+++ Localization/deDE.lua
@@ -1,9 +1,9 @@
-local addonName, addon = ...;
-local L = LibStub(addon.Libs.AceLocale):NewLocale(addonName, "deDE");
+local addonName, addon = ...
+local L = addon.Localization.GetLocale("deDE")
 if not L then return end
-addon.L = L;
+addon.L = L
 
-KrowiEVU.PluginsApi:LoadPluginLocalization(L);
+KrowiEVU.PluginsApi:LoadPluginLocalization(L)
 
 -- [[ https://legacy.curseforge.com/wow/addons/krowi-extended-vendor-ui/localization ]] --
 -- [[ Everything after this line is automatically generated from CurseForge and is not meant for manual edit - SOURCETOKEN - AUTOGENTOKEN ]] --